+++ mapnet/dev/get_doid_known_maps.py
@@ -1,21 +1,4 @@
-## Make a list of known mappings between doid and DOID from the owl files.
-
-# from indra.databases import owl_client
-
-# owl_clin = owl_client.OwlClient(prefix='doid')
-
-# res = owl_clin.update_from_file(prefix='doid', file = 'resources/doid.owl')
-
-
 from deeponto.onto import Ontology, OntologyPruner
-from deeponto.utils import (
-    InvertedIndex,
-    Tokenizer,
-    print_dict,
-    process_annotation_literal,
-    split_java_identifier,
-    uniqify,
-)
 
 doid_path = "resources/doid.owl"
 doid = Ontology(doid_path)
